Logging error on main error handler
