+++ src/app.js
@@ -8,7 +8,7 @@ const filesRouter = require('./middlewares/routers/files_router.js');
 const serversRouter = require('./middlewares/routers/servers_router.js');
 const usersRouter = require('./middlewares/routers/users_router.js');
 const ConnectionPoolFactory = require('./lib/factories/connection_pool_factory.js');
-const errorMiddleware = require('./middlewares/error_middleware.js');
+const errorMiddleware = require('./middlewares/main_error_handler.js');
 
 const logger = new Logger();
 const postgrePool = new ConnectionPoolFactory(logger).createPool();
@@ -27,7 +27,7 @@ serversRouter(app, logger, postgrePool);
 usersRouter(app, logger, postgrePool);
 
 // Add basic error middleware to handle all errors
-app.use(errorMiddleware);
+errorMiddleware(app, logger);
 
 // Start the app in the designated port and host
